Add an enhanced floating widget for tracking current status.
Note that the charm/intimidate tracking is very limited. The primary
problem is that there is currently no way to track the order things
happen in. These bonus points are only displayed for the user's benefit,
and aren't currently taken into account when disabling options, so the
user must be careful to consider how many charm or intimidate points
they will actually have at a given point in their playthrough.

Signed-off-by: Jason Lynch <jason@aexoden.com>