Rewrite dependency handling to allow for multiple required dependencies.
This will no doubt be insufficient when I get to a later option, but one
step at a time. The biggest remaining issue is that there is no
recursive check of dependencies, so a user could end up with an
impossible score if they select points on something, but then break a
dependency chain leading to that item.

For now, I'm simply copying over the recursive dependencies manually,
but this is potentially fragile.

Signed-off-by: Jason Lynch <jason@aexoden.com>