Implement a new algorithm for calculating available points and add a small test suite.
The previous algorithm for available points did not consider mutually
exclusive options, leading to incorrect totals. As far as I can tell,
the addition is now correct. However, I have uncovered some potential
issues with either the Wiki totals or my interpretation of the
dependencies that may need to be resolved.

Signed-off-by: Jason Lynch <jason@aexoden.com>