Add the remaining choices.
The overall available Paragon score matches what the wiki claims. The
Renegade score is 11 high, which could potentially be explained by the
one TODO item in the Renegade alignment assignment as well as the extra
2 points in Virmire that need to be researched.

The other inconsistency remains Noveria, which may just be a wiki
mistake, but it's hard to say.

Signed-off-by: Jason Lynch <jason@aexoden.com>